pyproject.toml
[project]
name = "mcp-server-docx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]==1.26.0",
    "python-docx==1.2.0",
]
